Add files via upload
this code is used to open each of the 3 chambers at different altitudes. the issue the has been encountered with the code is that the altimeter is reading a altitude of over 44,000 meters which it shouldn't, and its reading zero pressure. My guess is that the library I'm using for the altimeter is incorrect or the calculation for solving the altitude is in correct.